requirements.txt
ttkbootstrap>=1.6.1
GitPython>=3.1.30
requests>=2.28.2